Extracting magic multiple activation parser from physical jump damages. Adapting physical jump damages to handle the timed jump delay case. Adding healing turn split parser for magic effect 30.
